es/requirements.txt
elasticsearch
python-dotenv
tqdm
google-generativeai
ollama
openpyxl
python-telegram-bot
requests